refactor(test): extract cron issue-regression harness and frozen-time helper
(cherry picked from commit 47736e3432b7129cdbc9b0f631519bb35e1ece4c)
